requirements.txt
beautifulsoup4==4.9.3
ddt==1.4.1
msedge-selenium-tools==3.141.2
pytest==6.1.1
pytest-html==2.1.1
requests==2.24.0
selenium==3.141.0
